Reposition the backup interfaces definition to a more suitable location.
