+++ src/types/backup.ts
@@ -1,0 +1,10 @@
+export interface IBackUp {
+  name: string;
+  lastModified: string;
+  etag: string;
+  size: number;
+}
+
+export default interface IBackups {
+  backups: IBackUp[];
+}
